debug reset function

+++ index.js
@@ -161,8 +161,10 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
     var resetBtn = document.getElementById('resetBtn');
     resetBtn.addEventListener('click', function (event) {
         // if (isRun) return
-        km.delete();
-        km = null;
+        if (km !== null) {
+            km.delete();
+            km = null;
+        }
         isRun = false;
 
         // 清空畫布
